tests/PathSumIIITests.cs
using LeetCode.PathSumIII;

namespace tests;

public class PathSumIIITests
{
  // convert int array to binary tree
  private TreeNode ToTreeNode(int?[] nodes)
  {
    if (nodes == null || nodes.Length == 0 || nodes[0] == null) return null;

    TreeNode root = new TreeNode((int)nodes[0]);

    var queue = new Queue<TreeNode>();
    queue.Enqueue(root);
    int i = 1;  // tracking index of the nodes array
    while (queue.Any() && i < nodes.Length)
    {
      var node = queue.Dequeue();

      // add left node
      if (nodes[i] != null)
      {
        var left = new TreeNode((int)nodes[i]);
        node.left = left;
        queue.Enqueue(left);
      }
      // add right node
      if (i + 1 < nodes.Length && nodes[i + 1] != null)
      {
        var right = new TreeNode((int)nodes[i + 1]);
        node.right = right;
        queue.Enqueue(right);
      }
      i += 2;
    }
    return root;
  }

  public static IEnumerable<object[]> GetTestData()
  {
    yield return new object[]{
      new int?[]{10,5,-3,3,2,null,11,3,-2,null,1},
      8,
      3,
    };
    yield return new object[]{
      new int?[]{5,4,8,11,null,13,4,7,2,null,null,5,1},
      22,
      3
    };
  }

  [Theory]
  [MemberData(nameof(GetTestData))]
  public void Test1(int?[] nums, int targetSum, int expect)
  {
    var root = ToTreeNode(nums);
    Assert.Equal(expect, new Solution().PathSum(root, targetSum));
  }
}